Fix build

+++ source/MRMesh/MRSolidOfRevolution.cpp
@@ -12,7 +12,7 @@ Mesh makeSolidOfRevolution( const Contour2f& profile, int resolution )
     const auto profileSize = profile.size() - int( isClosed );
 
     std::vector<Vector3f> points( profileSize * resolution );
-    ParallelFor( points, [&] ( int index )
+    ParallelFor( points, [&] ( size_t index )
     {
         const auto angle = 2.f * PI_F * (float)( index / profileSize ) / (float)resolution;
         const auto x = std::cos( angle ), y = std::sin( angle );
